+++ src/lib/lofar_udp_structs_metadata.c
@@ -174,7 +174,7 @@ lofar_udp_metadata* lofar_udp_metadata_alloc() {
 
 	return meta;
 }
-sigproc_hdr* sigproc_hdr_alloc(int32_t fchannels) {
+sigproc_hdr* sigproc_hdr_alloc(uint64_t fchannels) {
 	DEFAULT_STRUCT_ALLOC(sigproc_hdr, hdr, sigproc_hdr_default, ;, NULL);
 
 	if (fchannels) {
